requirements.txt
# Core ML Libraries
torch==2.1.2
torchvision==0.16.2
numpy==1.26.3
pandas==2.1.4
scikit-learn==1.4.0
Pillow==10.2.0

# API Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Experiment Tracking
mlflow==2.10.0

# Data Versioning
dvc==3.38.1
dvc-s3==3.0.1

# Monitoring & Metrics
prometheus-client==0.19.0

# Testing
pytest==7.4.4
pytest-cov==4.1.0
pytest-asyncio==0.23.3
httpx==0.26.0

# Code Quality
ruff==0.1.14
black==24.1.1

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
tqdm==4.66.1
matplotlib==3.8.2
seaborn==0.13.1

# Logging
structlog==24.1.0

